Map File Clean Up
