+++ react-app/src/components/Maps/Locate.js
@@ -1,0 +1,22 @@
+const Locate = ({ panTo }) => {
+    return (
+      <button
+        className="locate"
+        onClick={() => {
+          navigator.geolocation.getCurrentPosition(
+            (position) => {
+              panTo({
+                lat: position.coords.latitude,
+                lng: position.coords.longitude,
+              });
+            },
+            () => null
+          );
+        }}
+      >
+        <img src="/compass.svg" alt="compass" />
+      </button>
+    );
+  }
+
+export default Locate;
