+++ react-app/src/components/SearchBar/SearchBar.js
@@ -1,9 +1,11 @@
 import React from "react";
 import "./SearchBar.css";
+// import Search from '../Maps/Search.js'
 
 const SearchBar = () => {
     return (
         <div className="search-wrapper">
+            {/* <Search /> */}
             <input type="text" placeholder="Enter a city" />
             <button type="submit">
                 <i className="fas fa-search"></i>
