+++ react-app/src/services/mapFunc.js
@@ -1,0 +1,25 @@
+import React from "react";
+import {
+  GoogleMap,
+  useLoadScript,
+  Marker,
+  InfoWindow,
+} from "@react-google-maps/api";
+import usePlacesAutocomplete, {
+  getGeocode,
+  getLatLng,
+} from "use-places-autocomplete";
+
+import usePlacesAutocomplete, {
+    getGeocode,
+    getLatLng,
+  } from "use-places-autocomplete";
+
+const convertAddressToCoords = async (address) => {
+      const results = await getGeocode({ address });
+      const { lat, lng } = await getLatLng(results[0]);
+      return {lat, lng}
+  };
+
+
+console.log(convertAddressToCoords('12521 Valley Pines Drive, Reisterstown, MD 21136, USA'))
